requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coverage==7.3.2
Flask==2.2.2
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
more-itertools==8.14.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.8
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
